pyproject.toml
[project]
name = "hamyo"
version = "0.1.0"
description = ""
authors = [
    {name = "DinzGIT",email = "112818583+ladinzgit@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite (>=0.21.0,<0.22.0)",
    "discord-py (>=2.5.2,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "jamo (>=0.4.1,<0.5.0)"
]

[tool.poetry]
packages = [{include = "hamyo", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
